Create slice for weather and reorganize mock data
