+++ src/features/quotes/Quote.js
@@ -2,17 +2,12 @@ import { selectQuote } from "./quoteSlice"
 import { useSelector } from "react-redux"
 
 export const Quote = () => {
-//const quote = useSelector(selectQuote)
-	const exampleQuote = {
-		quote: '"Every rose has its thorn."',
-		author: 'Bret Michaels',
-		category: 'Famous'
-	}
+	const quote = useSelector(selectQuote)
 
 	return  (
 		<div className="quote-container">
-			<p className="quote">{exampleQuote.quote}</p>
-			<p className="author">-{exampleQuote.author}</p>
+			<p className="quote">{quote.quote}</p>
+			<p className="author">{quote.author}</p>
 		</div>
 	)
