+++ src/features/quotes/quoteSlice.js
@@ -1,9 +1,16 @@
 import { createSlice } from "@reduxjs/toolkit";
 
+const exampleQuote = {
+		quote: '"Every rose has its thorn."',
+		author: 'Bret Michaels',
+		category: 'Famous'
+	}
+
+
 const quoteSlice = createSlice(
 	{
 		name: 'quote',
-		initialState: {},
+		initialState: exampleQuote,
 		reducers: {
 			getQuote: (state, action) => {
 				// api call?
